Add files via upload
Project Title: Hypothesis Testing on Cutlets Dataset
Description: This project involves applying hypothesis testing to the Cutlets dataset. I calculated the p-value to test the given hypotheses and determined whether to accept the null hypothesis or reject it in favor of the alternative hypothesis. This exercise demonstrates the practical application of statistical inference techniques in decision-making.